added sample variable, and changed to sampling_factor

+++ workflows/tasks/download.wdl
@@ -5,20 +5,23 @@ import "structs/compute.wdl"
 task srr {
     input {
         String srr
-        Float sampling_rate = 1
+        String sample = "false"
+        Float sampling_factor = 1
         Resources resources
     }
 
     command <<<
         prefetch ~{srr} --max-size 100G
         fasterq-dump -c 4000MB -b 4000MB -m 4000MB --split-files --threads 8  ~{srr}
-        seqtk sample -s 11 "~{srr}_1.fastq" ~{sampling_rate} > "~{srr}_1_sampled.fastq"
-        seqtk sample -s 11 "~{srr}_2.fastq" ~{sampling_rate} > "~{srr}_2_sampled.fastq"
-        rm -rf ~{srr}/
-        rm -rf "~{srr}_1.fastq"
-        rm -rf "~{srr}_2.fastq"
-        mv "~{srr}_1_sampled.fastq" "~{srr}_1.fastq"
-        mv "~{srr}_2_sampled.fastq" "~{srr}_2.fastq"
+        if [ "~{sample}" = "true" ]; then
+            seqtk sample -s 11 "~{srr}_1.fastq" ~{sampling_factor} > "~{srr}_1_sampled.fastq"
+            seqtk sample -s 11 "~{srr}_2.fastq" ~{sampling_factor} > "~{srr}_2_sampled.fastq"
+            rm -rf ~{srr}/
+            rm -rf "~{srr}_1.fastq"
+            rm -rf "~{srr}_2.fastq"
+            mv "~{srr}_1_sampled.fastq" "~{srr}_1.fastq"
+            mv "~{srr}_2_sampled.fastq" "~{srr}_2.fastq"
+        fi
     >>>
 
     output {
